moved commands out of extension into profile.ts
